+++ app/category/routers_category.py
@@ -1,11 +1,11 @@
 from fastapi import Depends, APIRouter
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from category.model_category import CategoryModel
+from category.models_category import CategoryModel
 from routers.consts import RouteSlug
 from models.database import get_session
-from category.schema_category import CategoryOptionalSchema
-from category.service_category import CategoryService
+from category.schemas_category import CategoryOptionalSchema
+from category.services_category import CategoryService
 from dependencies.base_dependency import (
     ValidGetByIdDependency,
     valid_empty_schema,
