+++ app/category/services_category.py
@@ -1,6 +1,6 @@
 from typing import Type
 
-from category.model_category import CategoryModel
+from category.models_category import CategoryModel
 from services.base_service import AbstractService
